Fix crash when syncing large IMAP mailboxes.
imap_sync_mailbox() and imap_exec_msgset() make a copy of the headers
before resorting them to generate a message-set, to avoid context
changes.  I noticed the oddity in the past but didn't investigate
deeply enough at the time. :-/

The code in imap_sync_mailbox() was wrong for four reasons:
1) Because IMAP_REOPEN_ALLOW is set, sync_helper() can trigger
   an imap_exec if the queue fills up, resulting in new messages
   being downloaded or expunges being triggered.

2) The copy was only allocating msgcount headers, instead of hdrmax.
   Downloading new messages could then attempt to append beyond the end
   of the array if hdrmax > msgcount.

3) New messages or expunges would disappear when the copy was restored
   afterwards.

4) The callers of mx_sync_mailbox() are prepared for context changes,
   and will adjust the cursor properly to avoid jumps.

One possible solution would be to turn off IMAP_REOPEN_ALLOW around
the calls to sync_helper(), but I've changed to sort and resort back
because I think making a copy is icky.

imap_exec_msgset() is another story.  Unfortunately, this is not
always called when IMAP_REOPEN_ALLOW is set (e.g. when server-side
copying messages), and the callers are not prepared to handle context
changes.  This function does make the copy if the reopen is not set.